Reorganizing more things
