+++ src/main/java/net/lomeli/lomlib/client/DevCapes.java
@@ -1,4 +1,4 @@
-package net.lomeli.lomlib.client.capes;
+package net.lomeli.lomlib.client;
 
 import java.awt.Graphics;
 import java.awt.Image;
