+++ src/main/java/net/lomeli/lomlib/client/ShapedFluidRecipeHandler.java
@@ -1,6 +1,5 @@
-package net.lomeli.lomlib.client.nei;
+package net.lomeli.lomlib.client;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import net.minecraft.item.ItemStack;
@@ -9,32 +8,33 @@ import net.minecraft.item.crafting.IRecipe;
 import net.minecraft.util.StatCollector;
 
 import net.lomeli.lomlib.libs.Strings;
-import net.lomeli.lomlib.recipes.ShapelessFluidRecipe;
+import net.lomeli.lomlib.recipes.ShapedFluidRecipe;
 
 import codechicken.core.ReflectionManager;
 import codechicken.nei.NEIServerUtils;
-import codechicken.nei.recipe.ShapelessRecipeHandler;
+import codechicken.nei.recipe.ShapedRecipeHandler;
 
-public class ShapelessFluidRecipeHandler extends ShapelessRecipeHandler {
+public class ShapedFluidRecipeHandler extends ShapedRecipeHandler {
 
     @Override
     public String getRecipeName() {
-        return StatCollector.translateToLocal(Strings.NEI_SHAPELESS);
+        return StatCollector.translateToLocal(Strings.NEI_SHAPED);
     }
 
     @SuppressWarnings("unchecked")
     @Override
     public void loadCraftingRecipes(String outputId, Object... results) {
-        if (outputId.equals("crafting") && getClass() == ShapelessFluidRecipeHandler.class) {
+        if (outputId.equals("crafting") && getClass() == ShapedFluidRecipeHandler.class) {
             List<IRecipe> allrecipes = CraftingManager.getInstance().getRecipeList();
             for (IRecipe irecipe : allrecipes) {
-                CachedShapelessRecipe recipe = null;
-                if (irecipe instanceof ShapelessFluidRecipe)
-                    recipe = fluidShapelessRecipe((ShapelessFluidRecipe) irecipe);
+                CachedShapedRecipe recipe = null;
+                if (irecipe instanceof ShapedFluidRecipe)
+                    recipe = fluidShapedRecipe((ShapedFluidRecipe) irecipe);
 
                 if (recipe == null)
                     continue;
 
+                recipe.computeVisuals();
                 arecipes.add(recipe);
             }
         }else {
@@ -48,13 +48,14 @@ public class ShapelessFluidRecipeHandler extends ShapelessRecipeHandler {
         List<IRecipe> allrecipes = CraftingManager.getInstance().getRecipeList();
         for (IRecipe irecipe : allrecipes) {
             if (NEIServerUtils.areStacksSameTypeCrafting(irecipe.getRecipeOutput(), result)) {
-                CachedShapelessRecipe recipe = null;
-                if (irecipe instanceof ShapelessFluidRecipe)
-                    recipe = fluidShapelessRecipe((ShapelessFluidRecipe) irecipe);
+                CachedShapedRecipe recipe = null;
+                if (irecipe instanceof ShapedFluidRecipe)
+                    recipe = fluidShapedRecipe((ShapedFluidRecipe) irecipe);
 
                 if (recipe == null)
                     continue;
 
+                recipe.computeVisuals();
                 arecipes.add(recipe);
             }
         }
@@ -65,13 +66,14 @@ public class ShapelessFluidRecipeHandler extends ShapelessRecipeHandler {
     public void loadUsageRecipes(ItemStack ingredient) {
         List<IRecipe> allrecipes = CraftingManager.getInstance().getRecipeList();
         for (IRecipe irecipe : allrecipes) {
-            CachedShapelessRecipe recipe = null;
-            if (irecipe instanceof ShapelessFluidRecipe)
-                recipe = fluidShapelessRecipe((ShapelessFluidRecipe) irecipe);
+            CachedShapedRecipe recipe = null;
+            if (irecipe instanceof ShapedFluidRecipe)
+                recipe = fluidShapedRecipe((ShapedFluidRecipe) irecipe);
 
-            if (recipe == null)
+            if (recipe == null || !recipe.contains(recipe.ingredients, ingredient))
                 continue;
 
+            recipe.computeVisuals();
             if (recipe.contains(recipe.ingredients, ingredient)) {
                 recipe.setIngredientPermutation(recipe.ingredients, ingredient);
                 arecipes.add(recipe);
@@ -79,20 +81,25 @@ public class ShapelessFluidRecipeHandler extends ShapelessRecipeHandler {
         }
     }
 
-    public CachedShapelessRecipe fluidShapelessRecipe(ShapelessFluidRecipe recipe) {
-        ArrayList<?> items;
+    public CachedShapedRecipe fluidShapedRecipe(ShapedFluidRecipe recipe) {
+        int width;
+        int height;
+        Object[] items;
         try {
-            items = ReflectionManager.getField(ShapelessFluidRecipe.class, ArrayList.class, recipe, 1);
+            width = ReflectionManager.getField(ShapedFluidRecipe.class, Integer.class, recipe, 4);
+            height = ReflectionManager.getField(ShapedFluidRecipe.class, Integer.class, recipe, 5);
+            items = ReflectionManager.getField(ShapedFluidRecipe.class, Object[].class, recipe, 3);
         }catch (Exception e) {
             e.printStackTrace();
             return null;
         }
 
-        for (int i = 0; i < items.size(); i++) {
-            if (items.get(i) instanceof List && ((List<?>) items.get(i)).isEmpty())
+        for (int i = 0; i < items.length; i++) {
+            if (items[i] instanceof List && ((List<?>) items[i]).isEmpty())
                 return null;
         }
 
-        return new CachedShapelessRecipe(items, recipe.getRecipeOutput());
+        return new CachedShapedRecipe(width, height, items, recipe.getRecipeOutput());
     }
+
 }
