+++ src/main/java/net/lomeli/lomlib/client/gui/element/TabTracker.java
@@ -1,4 +1,4 @@
-package net.lomeli.lomlib.client.gui.tab;
+package net.lomeli.lomlib.client.gui.element;
 
 public class TabTracker {
